1267 - start of caching in_taxons result via product groups
Checking in for future reference

Basic idea: each taxon has a product group which contains
a scope for calculating the products in a taxon AND its descendants.
We then rely on PGs to do the caching.

Seems to give a good speed-up (eliminating both taxon testing and
removal of duplicates). But the update of caching of taxons (both
for individual changes and for bulk operations) needs more work,
eg recalculate bottom-up.

Remark: this setup is useful for when we want to create taxons
dynamically from product groups, eg create a taxonomy of brands
based on information in a product's property list (ie using
product groups to collect together products of the same brand),
without needing to set this manually in the admin interface etc

[#1267 state:open]